chore: move mocks into `/test`
