+++ cmd/dnsapi/cloudflare_test.go
@@ -3,7 +3,7 @@ package dnsapi
 import (
 	"bytes"
 	"github.com/daruzero/cloudflare-dns-auto-updater-go/cmd/config"
-	"github.com/daruzero/cloudflare-dns-auto-updater-go/cmd/mocks"
+	"github.com/daruzero/cloudflare-dns-auto-updater-go/test/mocks"
 	"io"
 	"net/http"
 	"testing"
